+++ ArcdpsLogManager/Logs/Updates/LogDataUpdater.cs
@@ -201,6 +201,10 @@ namespace GW2Scratch.ArcdpsLogManager.Logs.Updates
 			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 1, 3)
 								 && log.Encounter == Encounter.Skorvald,
 				"Fix success detection for Skorvald when all players are dead while the boss is invulnerable at 1%"),
+			new LogUpdate(log => log.ParsingVersion < new Version(1, 11, 1, 4)
+			                     && log.Encounter == Encounter.Other
+								 && log.MapId == MapIds.LonelyTower,
+				"Added support for Eparch in the Lonely Tower fractal"),
 			// When adding a new update, you need to increase the revision (last value) of the version in the .csproj file
 			// unless the version changes more significantly, in that case it can be reset to 0.
 		};